bug(team): Add role check to create team button (#106)
REST api requires that user have 'editor' or 'admin' access to create a team.  Update UI to hide create button if user does not have either of those roles.
Ideally we would have used the AuthorizationDirective for this, but doing so would require moving it to a SharedModule since team is a submodule of core.