fix: header/footer banner overflow
If banner content is too long to fit within the banner it will overflow on top of the element below it.  This is due to the use of explicit heights.  These are a holdover from the old site layout that relied on fixed positions and margins.  They are no longer needed since #228.
Removing them will allo the banner to grow to contain the content.