fix: Add ViewportScroller implementation to properly handle scrolling on navigation events (#337)
Currently, scroll position doesn't always get reset properly when navigating routes.
Default angular behavior assumes that scrolling applies to the entire page.  When the primary scrolling viewport is a subelement (as it is here), a custom viewport scroller is needed.
Implementation is taking from the default `BrowserViewportScroller`, but operates on the configured element vs. the window.
Provider function provided for simply configuration.